refactor: Sprint 4 — decompose conversation-handler.ts (2021→602 LOC)
- Extract buildDynamicContext() to lib/voice/dynamic-context.ts
- Move 27 inline tools to 7 new IORS domain files:
  task-tools (3), mod-tools (3), integration-tools (2),
  planning-tools (4), memory-tools (3), skill-goal-tools (5),
  communication-tools (5)
- Replace 768 LOC executeTool() with 5-line thin wrapper
- Update IORS registry: 33 total tools across 11 domain files
- All 6 public exports unchanged (zero breaking changes)
